version 2.0.0 - fix decryption and add friendlier API
