+++ README.md
@@ -9,3 +9,23 @@ This is ported from CryptoJS.  The latest version can be found
 
 [crypto-js1]: https://github.com/brix/crypto-js
 [crypto-js2]: https://github.com/sytelus/CryptoJS
+
+
+## Example
+
+```dart
+import 'package:tripledes/tripledes.dart';
+
+main() {
+  var key = "cipher";
+  var blockCipher = new BlockCipher(new DESEngine(), key);
+  var message = "Driving in from the edge of town";
+  var ciphertext = blockCipher.encodeB64(message);
+  var decoded = blockCipher.decodeB64(ciphertext);
+
+  print("key: $key");
+  print("message: $message");
+  print("ciphertext (base64): $ciphertext");
+  print("decoded ciphertext: $decoded");
+}
+  ```
