+++ test/tripledes_test.dart
@@ -1,6 +1,7 @@
 // Copyright (c) 2017, john. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
+import 'package:tripledes/src/utils.dart';
 import 'package:tripledes/tripledes.dart';
 import 'package:test/test.dart';
 import 'dart:convert';
